+++ include/builders.hpp
@@ -2,9 +2,9 @@
 
 #include <array>
 #include <base_evaluator_factory.hpp>
-#include <concept_board_state_calculator_registry.hpp>
+#include <concepts/board_state_calculator_registry.hpp>
 #include <concept_composite_concepts.hpp>
-#include <concept_single_board_state_provider.hpp>
+#include <concepts/single_board_state_provider.hpp>
 #include <functional>
 #include <game_board_for_concepts.hpp>
 #include <memory>
