--- include/concept_board_state_tracker.hpp
@@ -1,15 +1,0 @@
-#pragma once
-
-#include <move_data_structs.hpp>
-#include <concepts>
-
-
-template <typename T>
-concept BoardStateTrackerConcept = requires(
-    T t,
-    const gameboard::ExecutedMove &executed_move
-) {
-    {t.UpdateBoardState(executed_move)} -> std::same_as<void>;
-};
-
-
