+++ include/concept_composite_concepts.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <concept_board_state_calculator.hpp>
-#include <concept_board_state_calculator_registry.hpp>
-#include <concept_single_board_state_provider.hpp>
+#include <concepts/board_state_calculator_registry.hpp>
+#include <concepts/single_board_state_provider.hpp>
 #include <concept_space_info_provider.hpp>
 #include <concepts>
