+++ include/concepts/board_state_calculator_registry.hpp
