+++ include/zobrist_for_concepts.hpp
@@ -5,7 +5,7 @@
 #include <board_data_structs.hpp>
 #include <chrono>
 #include <concept_multi_board_state_provider.hpp>
-#include <concept_single_board_state_provider.hpp>
+#include <concepts/single_board_state_provider.hpp>
 #include <evaluator_data_structs.hpp>
 #include <key_generator.hpp>
 #include <memory>
